+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
+## 1.2.0
+* feat: Support by switching language via the [BepInEx.ConfigurationManager](https://github.com/BepInEx/BepInEx.ConfigurationManager)
+
 ## 1.1.0
-* feat: Add shortcut keys to switch languages, requires manual configuration
+* feat: Add shortcut keys to switch languages, requires manual configuration first
 
 ## 1.0.2
-* fix: text on the map does not switch to the selected language
+* fix: Text on the map does not switch to the selected language